Restore integrated gas masks to military helmets.
This does unfortunately result in inflated fire protection values when
activated, due to how fire protection values are calculated differently
depending on the environmental_protection value of the armour. However,
while there's currently no way to work-around this (e.g. by overriding
fire protection explicitly instead of relying on material properties),
it doesn't have all that much of a real in-game impact, as in a fire
you'd still die to damage on your other body parts. Not having a gasmask
has much more of an in-game impact, and also doesn't make a lot sense.